Use normalised fitness instead of raw one

+++ include/StatsObserver.hpp
@@ -10,11 +10,11 @@ typedef std::lock_guard<std::mutex> lock;
 class ResultsContainer
 {
   private:
-    list<unsigned long long> fitness_list_;
+    list<double> fitness_list_;
     list<uint> generation_list_;
 
   public:
-    ResultsContainer(list<unsigned long long> fitness_list, list<uint> generation_list)
+    ResultsContainer(list<double> fitness_list, list<uint> generation_list)
         : fitness_list_(fitness_list)
         , generation_list_(generation_list)
     {
@@ -22,7 +22,7 @@ class ResultsContainer
     ResultsContainer(){};
 
     list<uint> getGenerations() { return generation_list_; }
-    list<unsigned long long> getFitnesses() { return fitness_list_; }
+    list<double> getFitnesses() { return fitness_list_; }
 };
 
 class StatsObserver : public EvolAlgObserver
@@ -30,7 +30,7 @@ class StatsObserver : public EvolAlgObserver
   private:
     EvolAlg *observed_EvolAlg_ = nullptr;
 
-    list<unsigned long long> fitness_list_;
+    list<double> fitness_list_;
     list<uint> generation_list_;
     std::mutex lists_mutex_;
 
@@ -43,7 +43,7 @@ class StatsObserver : public EvolAlgObserver
     virtual void update();
     void setObservedEvolAlg(EvolAlg &p);
 
-    list<unsigned long long> getFitnesses();
+    list<double> getFitnesses();
     list<uint> getGenerations();
     ResultsContainer getGenerationsAndFitness();
 };
